refactor: rename files with word DOM
