+++ src/dom_component/MoneyInput.js
@@ -1,4 +1,4 @@
-import { Event, QuerySelector } from '../constants/HTML';
+import { Event, QuerySelector } from '../constants/Dom';
 import { $ } from '../utils/DomUtils';
 
 class MoneyInput {
